Make resolved address classes private
