+++ src/neo4j/_work/summary.py
@@ -33,7 +33,7 @@ from .._meta import preview
 if t.TYPE_CHECKING:
     import typing_extensions as te
 
-    from ..addressing import Address
+    from .._addressing import Address
     from ..api import ServerInfo
 
     _T = t.TypeVar("_T")
